requirements.txt
Flask==3.0.0
pymongo==4.6.0
werkzeug==3.0.1
python-dotenv==1.0.0

